requirements.txt
numpy==2.2.0
pandas==2.2.3
catboost==1.2.8
selenium==4.31.0
scipy==1.15.2
xlrd==2.0.1
openpyxl
psycopg2-binary
webdriver-manager
pywin32
python-dotenv
thefuzz
google-genai
pyotp
sentry-sdk
pytest
pytest-mock